#38 [REFACTOR] SVG 아이콘 파일을 public/icons로 이동 및 import 경로 정리
